refactor: use only pages instead of layouts and divide in different files what is client to what is server side
