--- src/app/[locale]/(homepageDialogDeeplink)/action/become-member/layout.tsx
@@ -1,33 +1,0 @@
-import { UserActionType } from '@prisma/client'
-import { redirect, RedirectType } from 'next/navigation'
-
-import { HomepageDialogDeeplinkLayout } from '@/components/app/homepageDialogDeeplinkLayout'
-import { getAuthenticatedData } from '@/components/app/pageUserProfile/getAuthenticatedData'
-import { PageProps } from '@/types'
-import { setCallbackQueryString } from '@/utils/server/searchParams'
-import { USER_ACTION_DEEPLINK_MAP } from '@/utils/shared/urlsDeeplinkUserActions'
-
-import { SuccessState } from './sucessState'
-
-export const dynamic = 'force-dynamic'
-
-export default async function Layout({ params, children }: React.PropsWithChildren<PageProps>) {
-  const { locale } = params
-  const user = await getAuthenticatedData()
-
-  if (!user) {
-    redirect(
-      USER_ACTION_DEEPLINK_MAP[UserActionType.OPT_IN].getDeeplinkUrl({
-        locale,
-        queryString: setCallbackQueryString({ destination: 'becomeMember' }),
-      }),
-      RedirectType.replace,
-    )
-  }
-
-  return (
-    <HomepageDialogDeeplinkLayout pageParams={params} size="sm">
-      {user.hasOptedInToMembership ? <SuccessState /> : children}
-    </HomepageDialogDeeplinkLayout>
-  )
-}
